generator

+++ kattisproblem/output_format_validators/validate.py
@@ -5,7 +5,7 @@ while True:
     try:
         out_ = input()
         count += 1
-        assert out_ == "Impossible" or 0 <= int(out_) <= 1000
+        assert out_ == "Impossible" or 0 <= float(out_) <= 1000
     except EOFError:
         break
 assert count > 0
